requirements.txt
asyncio
black
datalake-scripts
datetime
ipaddress
pytest
pytest-asyncio
python-dotenv
stix2
wheel